Added new difficulty "Difficulty Scalar"
This mod aims to add more customization to difficulty. The simplest way to think about it if hardrock or ez mod could be adjusted, meaning this mod only changes the multiplies.

It also has variables cap, hr is capped at ar 10, but with this mod it goes to 11.

The different is that if a mod has settings with 0, it defaults to using 0.1 * multiplier